# Refactor connection handling