This is an attempt at refactoring `Connection`  and `ConnectionHandler` to use `poll` instead of `async` `.await` syntax.

Advantages:

1. Ability to continue reading while in the middle of writing: previously a write operation would block read operations from happening. Now as soon as an operation returns `Poll::Pending` the other operations can be polled
2. Ability to continue reading and writing while in the middle of flushing: previously flushing blocked everything else until it completed
3. Easier control of connection state and reduced error branches because of centralized connection error handling
4. Removal of `select!` from a very critical path. The issues with it are the following:
    * It relies on the `Future`s polled by it's branches to be _cancel safe_. Cancel safety requires `Drop`ping of a `Future` to be a no-op. While this seemed to be true it's easy to get wrong.
    * Execution of the expression blocks all other operations (see point 1 and 2)
5. No more flush interval
6. Higher write performance thanks to flattening of small writes (buffering them ourselves) and vectored writes when possible

Left to implement:

* [x] Flush interval
* [x] Vectored writes
* [x] Disconnect handling
* [x] Write backpressure
* [x] Reimplement `flush`
* [x] Cleanup commits
* [x] Benchmarks

Fixes #905
Fixes #923
Fixes #582 
Closes #869
Closes #935
Closes #971 
Closes #1070