@paolobarbolini 2023-07-29
Benchmarks as of now:

<details>
    <summary>main</summary>

```console
async-nats: publish throughput/32
                        time:   [202.71 µs 208.83 µs 215.22 µs]
                        thrpt:  [14.180 MiB/s 14.614 MiB/s 15.055 MiB/s]
async-nats: publish throughput/1024
                        time:   [331.07 µs 343.49 µs 356.15 µs]
                        thrpt:  [274.20 MiB/s 284.30 MiB/s 294.97 MiB/s]
Found 1 outliers among 30 measurements (3.33%)
  1 (3.33%) high mild
async-nats: publish throughput/8192
                        time:   [2.2131 ms 2.3912 ms 2.5751 ms]
                        thrpt:  [303.39 MiB/s 326.72 MiB/s 353.01 MiB/s]

async-nats: publish messages amount/32
                        time:   [200.31 µs 205.57 µs 211.80 µs]
                        thrpt:  [472.13 Kelem/s 486.45 Kelem/s 499.22 Kelem/s]
Found 2 outliers among 30 measurements (6.67%)
  2 (6.67%) high mild
async-nats: publish messages amount/1024
                        time:   [325.23 µs 335.40 µs 346.18 µs]
                        thrpt:  [288.86 Kelem/s 298.15 Kelem/s 307.48 Kelem/s]
Found 2 outliers among 30 measurements (6.67%)
  1 (3.33%) high mild
  1 (3.33%) high severe
async-nats: publish messages amount/8192
                        time:   [1.9831 ms 2.2042 ms 2.4116 ms]
                        thrpt:  [41.466 Kelem/s 45.367 Kelem/s 50.427 Kelem/s]

subscribe amount/32     time:   [1.5607 ms 1.6374 ms 1.6978 ms]
                        thrpt:  [58.900 Kelem/s 61.074 Kelem/s 64.075 Kelem/s]
subscribe amount/1024   time:   [1.1902 ms 1.2250 ms 1.2595 ms]
                        thrpt:  [79.400 Kelem/s 81.633 Kelem/s 84.018 Kelem/s]
subscribe amount/8192   time:   [4.2955 ms 4.3576 ms 4.4296 ms]
                        thrpt:  [22.575 Kelem/s 22.949 Kelem/s 23.280 Kelem/s]
Found 2 outliers among 30 measurements (6.67%)
  2 (6.67%) high severe
```
</details>

<details>
    <summary>this branch</summary>

```console
async-nats: publish throughput/32
                        time:   [190.19 µs 196.71 µs 203.55 µs]
                        thrpt:  [14.992 MiB/s 15.514 MiB/s 16.045 MiB/s]
                 change:
                        time:   [-8.7393% -4.7959% -0.8225%] (p = 0.03 < 0.05)
                        thrpt:  [+0.8294% +5.0375% +9.5762%]
                        Change within noise threshold.
Found 1 outliers among 30 measurements (3.33%)
  1 (3.33%) high mild
async-nats: publish throughput/1024
                        time:   [308.62 µs 318.89 µs 329.49 µs]
                        thrpt:  [296.38 MiB/s 306.23 MiB/s 316.43 MiB/s]
                 change:
                        time:   [-8.6554% -2.5471% +3.8930%] (p = 0.46 > 0.05)
                        thrpt:  [-3.7471% +2.6137% +9.4756%]
                        No change in performance detected.
Found 2 outliers among 30 measurements (6.67%)
  1 (3.33%) high mild
  1 (3.33%) high severe
async-nats: publish throughput/8192
                        time:   [1.4773 ms 1.5160 ms 1.5551 ms]
                        thrpt:  [502.36 MiB/s 515.34 MiB/s 528.85 MiB/s]
                 change:
                        time:   [-37.950% -31.428% -24.041%] (p = 0.00 < 0.05)
                        thrpt:  [+31.650% +45.833% +61.161%]
                        Performance has improved.
Found 3 outliers among 30 measurements (10.00%)
  2 (6.67%) high mild
  1 (3.33%) high severe

async-nats: publish messages amount/32
                        time:   [192.45 µs 197.46 µs 202.29 µs]
                        thrpt:  [494.35 Kelem/s 506.44 Kelem/s 519.61 Kelem/s]
                 change:
                        time:   [-11.902% -7.2112% -2.1045%] (p = 0.01 < 0.05)
                        thrpt:  [+2.1497% +7.7717% +13.510%]
                        Performance has improved.
Found 1 outliers among 30 measurements (3.33%)
  1 (3.33%) high mild
async-nats: publish messages amount/1024
                        time:   [309.51 µs 318.95 µs 330.52 µs]
                        thrpt:  [302.55 Kelem/s 313.53 Kelem/s 323.09 Kelem/s]
                 change:
                        time:   [-10.980% -3.2048% +4.6338%] (p = 0.45 > 0.05)
                        thrpt:  [-4.4286% +3.3109% +12.334%]
                        No change in performance detected.
Found 3 outliers among 30 measurements (10.00%)
  1 (3.33%) high mild
  2 (6.67%) high severe
async-nats: publish messages amount/8192
                        time:   [1.4486 ms 1.4860 ms 1.5305 ms]
                        thrpt:  [65.339 Kelem/s 67.293 Kelem/s 69.030 Kelem/s]
                 change:
                        time:   [-31.125% -23.573% -15.210%] (p = 0.00 < 0.05)
                        thrpt:  [+17.938% +30.843% +45.191%]
                        Performance has improved.
Found 3 outliers among 30 measurements (10.00%)
  1 (3.33%) high mild
  2 (6.67%) high severe

subscribe amount/32     time:   [4.3207 ms 4.4703 ms 4.6061 ms]
                        thrpt:  [21.710 Kelem/s 22.370 Kelem/s 23.144 Kelem/s]
                 change:
                        time:   [+173.31% +198.66% +227.86%] (p = 0.00 < 0.05)
                        thrpt:  [-69.499% -66.517% -63.412%]
                        Performance has regressed.
Found 3 outliers among 30 measurements (10.00%)
  1 (3.33%) low mild
  2 (6.67%) high mild
subscribe amount/1024   time:   [1.3827 ms 1.4205 ms 1.4670 ms]
                        thrpt:  [68.166 Kelem/s 70.398 Kelem/s 72.321 Kelem/s]
                 change:
                        time:   [+14.539% +18.584% +23.016%] (p = 0.00 < 0.05)
                        thrpt:  [-18.710% -15.672% -12.694%]
                        Performance has regressed.
Found 2 outliers among 30 measurements (6.67%)
  2 (6.67%) high mild
subscribe amount/8192   time:   [4.2687 ms 4.3939 ms 4.5255 ms]
                        thrpt:  [22.097 Kelem/s 22.759 Kelem/s 23.426 Kelem/s]
                 change:
                        time:   [-3.3419% +3.2701% +9.7699%] (p = 0.35 > 0.05)
                        thrpt:  [-8.9003% -3.1665% +3.4574%]
                        No change in performance detected.
Found 2 outliers among 30 measurements (6.67%)
  2 (6.67%) high mild
```
</details>

I'll look into what's happening with subscriptions