@paolobarbolini 2023-07-30
I tried implementing vectored writes to see how hard it would be. It took very little. The results are awesome. I'll send them once this is merged though :smiley: 

![image](https://github.com/nats-io/nats.rs/assets/6215781/26174882-c72d-4ce1-8751-59c3448f69de)