@paolobarbolini 2023-07-30
There's a problem I didn't take into account before: calling `Client::flush()` now means waiting both for past **and future** writes to be flushed, while in the past it only acted on past writes. Think makes calling `flush` a footgun in some cases.

I'm kind of feeling like we could live without it. When it comes to SQL databases for examples I don't think I ever needed to flush the connection manually, so why should that apply for NATS? The real thing missing in nats.rs to make this a thing of the past is waiting for acknowledgments from the Core NATS protocol. Right now calling `subscribe` resolves as soon as the `Command` was written to the channel. What should instead happen is having it wait for the server to reply `+OK`. I think I can implement that in a follow-up PR.