@caspervonb 2023-07-31
> What should instead happen is having it wait for the server to reply +OK
+OK is only when verbose mode is on, and it isn't tied to any particular request 🤔