@paolobarbolini 2023-07-31
> > What should instead happen is having it wait for the server to reply +OK
> 
> +OK is only when verbose mode is on, and it isn't tied to any particular request thinking

Ah. I wasn't expecting that :smile:. So best case scenario we know when a command was written to the buffer. Nothing more :disappointed: