@Jarema 2023-08-01
@paolobarbolini I just did a very simple test:

```rust
    #[tokio::test]
    async fn request_bench() {
        use futures::stream::StreamExt;

        let server = nats_server::run_basic_server();

        let (tx, rx) = tokio::sync::oneshot::channel();

        tokio::spawn({
            let url = server.client_url().clone();
            async move {
                let client = async_nats::connect(url).await.unwrap();
                let mut subscription = client.subscribe("request".into()).await.unwrap();
                client.flush().await.unwrap();
                tx.send(()).unwrap();

                while let Some(message) = subscription.next().await {
                    client
                        .publish(message.reply.unwrap(), "".into())
                        .await
                        .unwrap();
                    client.flush().await.unwrap();
                }
            }
        });

        let client = async_nats::connect(server.client_url()).await.unwrap();
        rx.await.unwrap();

        let total = std::time::Instant::now();
        for _ in 0..100 {
            let now = std::time::Instant::now();
            let request = client.request("request".into(), "".into()).await.unwrap();
            let elapsed = now.elapsed();
            println!("request took: {:?}", elapsed);
        }
        println!("took: {:?}", total.elapsed());
    }
```

and run it against your branch and main.

**main**
```
running 1 test
request took: 386.5µs
request took: 347.334µs
request took: 325.792µs
request took: 312.375µs
request took: 401.667µs
request took: 293.042µs
request took: 267.291µs
request took: 289.584µs
request took: 311.125µs
request took: 278.708µs
request took: 307.666µs
request took: 305.625µs
request took: 282.375µs
request took: 311.5µs
request took: 257.459µs
request took: 211.375µs
request took: 225.083µs
request took: 238.333µs
request took: 235.292µs
request took: 220.166µs
request took: 309.167µs
request took: 347.625µs
request took: 246.208µs
request took: 220.083µs
request took: 275.166µs
request took: 208.542µs
request took: 201.75µs
request took: 242.667µs
request took: 289.167µs
request took: 258.583µs
request took: 243.5µs
request took: 236.458µs
request took: 235.75µs
request took: 213.5µs
request took: 226.167µs
request took: 201.708µs
request took: 210.541µs
request took: 241.709µs
request took: 213.209µs
request took: 213.083µs
request took: 218.209µs
request took: 216.166µs
request took: 231.25µs
request took: 226.25µs
request took: 201.416µs
request took: 225.083µs
request took: 214.5µs
request took: 154.666µs
request took: 149.334µs
request took: 146.875µs
request took: 199.917µs
request took: 158.333µs
request took: 165.792µs
request took: 200.083µs
request took: 150.042µs
request took: 139.167µs
request took: 158µs
request took: 159.833µs
request took: 131.25µs
request took: 170.167µs
request took: 165.667µs
request took: 146.75µs
request took: 170.625µs
request took: 175.083µs
request took: 139.209µs
request took: 154.25µs
request took: 128.458µs
request took: 130.458µs
request took: 155.042µs
request took: 154.042µs
request took: 151.75µs
request took: 154µs
request took: 152.708µs
request took: 200.208µs
request took: 186.5µs
request took: 179.542µs
request took: 252.083µs
request took: 194.375µs
request took: 190.709µs
request took: 177.625µs
request took: 186.875µs
request took: 191.875µs
request took: 206.209µs
request took: 179.75µs
request took: 201.541µs
request took: 181.292µs
request took: 188.292µs
request took: 183.333µs
request took: 191.917µs
request took: 182.417µs
request took: 190.667µs
request took: 181.708µs
request took: 182.834µs
request took: 233.792µs
request took: 267.417µs
request took: 195.459µs
request took: 216.958µs
request took: 203.333µs
request took: 220.083µs
request took: 199.917µs
took: 22.05625ms
```

**this PR**
```rust
running 1 test
request took: 1.636792ms
request took: 2.731041ms
request took: 1.58475ms
request took: 1.566958ms
request took: 2.643208ms
request took: 1.454792ms
request took: 2.466917ms
request took: 1.493792ms
request took: 2.574417ms
request took: 1.496417ms
request took: 2.665875ms
request took: 1.534542ms
request took: 2.622291ms
request took: 2.609041ms
request took: 1.47ms
request took: 2.647458ms
request took: 1.451333ms
request took: 2.544708ms
request took: 2.532542ms
request took: 2.517083ms
request took: 2.507375ms
request took: 2.603041ms
request took: 2.570875ms
request took: 1.361167ms
request took: 2.481333ms
request took: 2.571875ms
request took: 2.548333ms
request took: 2.52175ms
request took: 2.557666ms
request took: 1.356833ms
request took: 2.445416ms
request took: 2.476167ms
request took: 2.543875ms
request took: 2.538417ms
request took: 2.460625ms
request took: 2.619833ms
request took: 2.529833ms
request took: 1.352584ms
request took: 2.525375ms
request took: 2.547ms
request took: 2.559417ms
request took: 1.445042ms
request took: 2.46875ms
request took: 1.389834ms
request took: 2.439042ms
request took: 2.509292ms
request took: 2.513542ms
request took: 2.538291ms
request took: 2.528375ms
request took: 2.490459ms
request took: 2.564542ms
request took: 1.363833ms
request took: 2.516792ms
request took: 2.544042ms
request took: 2.559834ms
request took: 1.405125ms
request took: 2.496167ms
request took: 2.517917ms
request took: 2.500666ms
request took: 2.473292ms
request took: 2.540166ms
request took: 2.5195ms
request took: 2.458042ms
request took: 2.512125ms
request took: 2.523209ms
request took: 1.33225ms
request took: 2.49025ms
request took: 2.469333ms
request took: 2.486875ms
request took: 2.546042ms
request took: 2.488375ms
request took: 2.522208ms
request took: 2.511334ms
request took: 2.491667ms
request took: 2.538208ms
request took: 2.559042ms
request took: 2.411417ms
request took: 2.527666ms
request took: 2.467333ms
request took: 2.443625ms
request took: 2.483ms
request took: 2.47725ms
request took: 2.533583ms
request took: 2.491792ms
request took: 2.444958ms
request took: 2.4395ms
request took: 2.453959ms
request took: 2.480542ms
request took: 2.507917ms
request took: 2.518084ms
request took: 2.484042ms
request took: 2.501208ms
request took: 2.491292ms
request took: 2.497958ms
request took: 2.494875ms
request took: 2.515584ms
request took: 2.519917ms
request took: 2.520375ms
request took: 2.465583ms
request took: 2.533583ms
took: 234.451083ms
```

Removing the flush from publish on response makes them both as slow as this PR code.
I delierately used two client connections.

This is a huge difference.

So, we are well aware of the fact that there are optimizations to flushing/writing mechanism possible, but this does not seem to be the way.

Also note, that most, if not all NATS clients have manual flush option for exactly this reason.
Maybe we can be smarter, but we're not as of yet ;).